conclusão do nativo e inativo, agora conclusão cadastro ir pro show
