+++ public/js/cadastro.js
@@ -6,3 +6,16 @@ numerosInput.addEventListener("input", function() {
   }
 });
 
+function formatarValor() {
+  // Obtém o valor digitado no input
+  let valor = document.getElementById('valor').value;
+
+  // Remove todos os caracteres não numéricos (exceto ponto decimal, se houver)
+  valor = valor.replace(/\D/g, '');
+
+  // Formata o valor como dinheiro (adicionando ponto decimal e vírgula)
+  valor = (parseFloat(valor) / 100).toFixed(2).replace('.', ',');
+
+  // Atualiza o valor formatado no input
+  document.getElementById('valor').value = valor;
+}
